Update to add isHost to session member info
